refactor: :recycle: adding hexagonal architecture and repository interface
