+++ src/domain/enums/enumPaymentType.ts
