+++ src/domain/interfaces/IAccountOperations.ts
@@ -1,0 +1,8 @@
+import { Account } from "src/domain/entities/entitieAccount";
+
+export interface IAccountOperations {
+  deposit(amount: number): void;
+  checkBalance(): void;
+  withdraw(amount: number): void;
+  transfer(amount: number, toAccount: Account): void;
+}
