+++ src/domain/interfaces/IPayment.ts
@@ -1,0 +1,7 @@
+import { PaymentType } from "src/domain/enums/enumPaymentType"
+
+export interface IPayment {
+ type: PaymentType
+
+ pay(amountPayment: number, paymentReference: string): void
+}
