+++ src/domain/interfaces/IRepository.ts
@@ -1,0 +1,6 @@
+export interface IRepository<T> {
+ findAll(): Promise<T[]>;
+ findById(id: string): Promise<T | null>;
+ save(entity: T): Promise<T>;
+ delete(id: string): Promise<boolean>;
+}
