--- src/interfaces/Payment.ts
@@ -1,7 +1,0 @@
-import { PaymentType } from "src/enums/enumPaymentType"
-
-export interface Payment{
- type: PaymentType
-
- pay(amountPayment: number, paymentReference: string ): void
-}
