+++ src/modules/accountModule.ts
@@ -1,16 +1,16 @@
 import { Module, forwardRef } from '@nestjs/common';
-import { AccountService } from '../services/accountService';
-import { AccountController } from '../controllers/accountController';
+import { AccountService } from '../domain/services/accountService';
+import { AccountController } from '../application/controllers/accountController';
 import { CustomerModule } from './customerModule';
-import { ConcreteAccountFactory } from 'src/factorys/factoryAccount';
+import { ConcreteAccountFactory } from 'src/domain/factorys/factoryAccount';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Account } from 'src/entities/entitieAccount';
-import { AccountRepository } from 'src/repository/accountRepository';
+import { Account } from 'src/domain/entities/entitieAccount';
+import { AccountRepository } from 'src/infrastructure/repository/accountRepository';
 
 @Module({
-  imports: [forwardRef(() => CustomerModule), TypeOrmModule.forFeature([Account])], 
+  imports: [forwardRef(() => CustomerModule), TypeOrmModule.forFeature([Account])],
   providers: [AccountService, ConcreteAccountFactory, AccountRepository],
   controllers: [AccountController],
-  exports: [AccountService, ConcreteAccountFactory, AccountRepository], 
+  exports: [AccountService, ConcreteAccountFactory, AccountRepository],
 })
-export class AccountModule {}
+export class AccountModule { }
